[clang-tidy] Validate fixes in header files
The -header-files argument to check_clang_tidy is a comma
separted list of filenames to process from the input tree
into the output tree for inclusion by a checked source file.

The check_clang_tidy invocation in the source file is assumed
to add '.' to the include path in order to locate sanitized
headers with the CHECK prefixes removed.  It is also assumed
that the necessary options to clang-tidy will be present to
instruct clang-tidy to apply fixes to the corresponding headers.

Messages, fixes and notes are checked for each header listed.