[clang-tidy] Enhance run-clang-tidy.py to allow clang-tidy to fix/format
If you are only processing source files, and not applying fixits to
header files, having clang-tidy apply and format the fixes can be
faster than clang-apply-replacements.

1) clang-tidy is running in parallel on all the source files, while
   clang-apply-replacements is processing source files serially.
2) clang-tidy has already loaded the source into memory for analysis,
   so we might as well apply and format the fixes while we've got it
   loaded.

There is one caveat to the above, however.  If you are applying fixes to
header files with -header-filter, then while processing independent
source files they can both want to apply the same fix to the same
(shared) header file.  In that scenario, you probably want to use
clang-apply-replacements which will deduplicate the fixes.