[clang-tidy] Ignore macros expanding in contexts where enums can't be used
A macro considered for conversion to an enum might be used in a
context where an enum would be invalid.  For instance:

  #define BAD (1, 2)

  extern void f(int, int);

  void g()
  {
    f BAD;
  }

Here the parentheses in the macro definition aren't just used for
grouping an application of operator, but instead represent part of the
syntax of a function call.  Converting such a macro to an enum would
result in invalid code.

Record all the expansion locations of candidates for enum conversion and
match the expansion ranges to elements in the AST with a visitor.  If
the expansion range isn't completely within an expression AST node, then
discard the candidate.