[clang-tidy] Enable selective building of check modules
The CMake variable CLANG_TIDY_ENABLE_MODULES is set to either 'all' (the
default), or a semi-colon separated list of short names of check modules
to build.  The short name of a module is the same as the name of its
corresponding subdirectory.  Only named modules will have targets
created for them and their subdirectories added.